# docs/swagger.yaml
basePath: /
consumes:
- application/x-www-form-urlencoded
definitions:
...
  Error:
    properties:
      code:
        type: string
        x-go-name: Code
      message:
        type: string
        x-go-name: Message
    type: object
    x-go-package: organizr/server/types
  Task:
    properties:
      description:
        type: string
        x-go-name: Description
      gid:
        type: string
        x-go-name: Gid
      id:
        format: int64
        type: integer
        x-go-name: Id
      task_column_id:
        format: int64
        type: integer
        x-go-name: TaskColumnId
      title:
        type: string
        x-go-name: Title
    type: object
    x-go-package: organizr/server/types
...
host: localhost:1323
info:
  description: API Documentation.
  title: Organizr
  version: 1.0.0
paths:
...
  /api/task/{Column_GID}:
    get:
      description: Retrieves all tasks by parent column UUID
      operationId: task-retrieve-all
      parameters:
      - description: UUID of parent column
        in: path
        name: Column_GID
        required: true
        type: string
      responses:
        "200":
          $ref: '#/responses/multi-task-response'
        "400":
          $ref: '#/responses/error-response'
      security:
      - Bearer:
        - '[]'
      tags:
      - task
    post:
      description: Creates a task in the column specified by UUID
      operationId: task-create
      responses:
        "200":
          $ref: '#/responses/single-task-response'
        "400":
          $ref: '#/responses/error-response'
      security:
      - Bearer:
        - '[]'
      tags:
      - task
  /api/task/{Column_GID}/{Task_GID}:
    delete:
      description: Deletes task by parent column and task UUIDs
      operationId: task-delete
      parameters:
      - description: UUID of parent column
        in: path
        name: Column_GID
        required: true
        type: string
      - description: UUID of task
        in: path
        name: Task_GID
        required: true
        type: string
      responses:
        "200":
          description: ""
        "400":
          $ref: '#/responses/error-response'
      security:
      - Bearer:
        - '[]'
      tags:
      - task
    get:
      description: Retrieves task by parent column and task UUIDs
      operationId: task-retrieve-one
      parameters:
      - description: UUID of parent column
        in: path
        name: Column_GID
        required: true
        type: string
      - description: UUID of task
        in: path
        name: Task_GID
        required: true
        type: string
      responses:
        "200":
          $ref: '#/responses/single-task-response'
        "400":
          $ref: '#/responses/error-response'
      security:
      - Bearer:
        - '[]'
      tags:
      - task
    put:
      description: Updates task by parent column and task UUIDs
      operationId: task-update
      parameters:
      - description: UUID of parent column
        in: path
        name: Column_GID
        required: true
        type: string
      - description: UUID of task
        in: path
        name: Task_GID
        required: true
        type: string
      - description: Title of task
        in: body
        name: Title
        schema:
          type: string
      - description: Description of task
        in: body
        name: Description
        schema:
          type: string
      responses:
        "200":
          $ref: '#/responses/single-task-response'
        "400":
          $ref: '#/responses/error-response'
      security:
      - Bearer:
        - '[]'
      tags:
      - task
produces:
- application/json
responses:
  error-response:
    description: ""
    schema:
      $ref: '#/definitions/Error'
...
  multi-task-response:
    description: ""
    schema:
      items:
        $ref: '#/definitions/Task'
      type: array
...
  single-task-response:
    description: ""
    schema:
      $ref: '#/definitions/Task'
schemes:
- http
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"